requirements.txt
pyserial
dearpygui
